requirements.txt
arabic-reshaper==3.0.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
decorator==4.4.2
dnspython==2.3.0
email-validator==2.0.0.post2
ffmpeg==1.4
Flask==2.3.2
Flask-Admin==1.6.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
googletrans==3.0.0
idna==2.5
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
moviepy==1.0.3
numpy==1.25.0
Pillow==9.5.0
proglog==0.1.10
python-bidi==0.4.2
requests==2.31.0
SQLAlchemy==2.0.16
stability-sdk==0.8.3
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.3.6
WTForms==3.0.1
zipp==3.15.0
httplib2~=0.22.0
oauth2client~=4.1.3
google-api-python-client~=2.90.0
psycopg2-binary==2.9.9

